+++ src/Persistence/StatsDAO.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public interface StatsDAO {
-    public void insertStats(int gameId, double time, int num_coffees);
+    public void insertStats(int gameId, double time, double num_coffees);
     public List<CoffeeStats> loadStats();
     public void deleteAll();
     public List<Integer> loadAllPlayers();
